+++ frontend/src/lib/library-file-column-widths.ts
@@ -10,6 +10,7 @@ const LIBRARY_FILE_COLUMN_WIDTH_KEYS: MediaFileSortKey[] = [
   "hdr_type",
   "duration",
   "audio_codecs",
+  "audio_spatial_profiles",
   "audio_languages",
   "subtitle_languages",
   "subtitle_codecs",